+++ discord/cogs/backup/__init__.py
@@ -1,0 +1,5 @@
+from .backup import BackupCog
+
+
+async def setup(bot):
+    await bot.add_cog(BackupCog())